Animation Editor(7)
- 여러 장의 텍스쳐를 선택해서 여러 프레임을 한 번에 추가할 수 있도록 개선

+++ CButtonUI.cpp
@@ -35,10 +35,10 @@ void CButtonUI::Render_DerivedUI()
 		AlphaBlend(BackDC
 			, (int)(vFinalPos.x)
 			, (int)(vFinalPos.y)
-			, vScale.x, vScale.y
+			, (int)vScale.x, (int)vScale.y
 			, pTex->GetDC()
 			, 0, 0
-			, vScale.x, vScale.y
+			, (int)vScale.x, (int)vScale.y
 			, blend);
 	}
 	else
